### 2016-08-01 #61 @jsta
dbhydroR package
`dbhydroR` provides scripted access to the South Florida Water Management District's [DBHYDRO database](http://my.sfwmd.gov/dbhydroplsql/show_dbkey_info.main_menu) which holds over 35 million water quality and hydrologic data records from the Florida Everglades and surrounding areas.

```
Package: dbhydroR
Title: 'DBHYDRO' hydrologic and water quality data from R
Description: Client for programmatic access to the South Florida Water
  Management District's 'DBHYDRO' database at
  http://my.sfwmd.gov/dbhydroplsql/show_dbkey_info.main_menu, with functions
  for accessing hydrologic and water quality data.
Version: 0.1-6
Author: Joseph Stachelek <stachel2@msu.edu>
Maintainer: Joseph Stachelek <stachel2@msu.edu>
    httr,
    reshape2,
    RCurl,
    XML
License: GPL
```

https://github.com/sfwmd/dbhydro

Anyone interested in water quality and hydrologic data from the Florida Everglades - ecologists, engineers, meteorologists, hydrogeologists, hydrologists, etc.

For the most part, no. However, a very small proportion of DBHYDRO mirrors data that is also present on the USGS (United States Geological Survey) NWIS (National Water Information System). This data could then be accessed by the `dataRetrieval` package.

    - [ ] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.

Some components like the rOpenSci footer, README section ordering, and NEWS tags are pending this initial inquiry.
