### 2016-08-01 #61 @sckott
thanks for your submission @jsta  - looks like you didn't check _Do you intend for this package to go on CRAN?_. Can you check it or explain why not
