### 2016-08-01 #61 @sckott
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)
- [ ] **Automated tests:** __Package has a test suite but all tests are skipped__

---

#### Editor comments

Looks like your package has a test suite, but all tests are skipped on Travis. Why is that? They do run locally for me, which is good.
