### 2016-08-01 #61 @jsta
After working with the `rnoaa` package I was under the impression that this was the norm.
 * https://github.com/ropensci/rnoaa/blob/master/tests/testthat/test-buoy.R
 * https://github.com/ropensci/rnoaa/blob/master/tests/testthat/test-coops.R

Should I enable them all on Travis or only certain types?
