### 2016-08-02 #61 @sckott
@jsta sorry about misleading!  I don't remember exactly why but I usually only skip on travis temporarily if a web service is down temporarily .

> Should I enable them all on Travis or only certain types?

Just skip on CRAN, and enable in all others
