### 2016-08-02 #61 @jsta
Ok, tests are enabled on Travis now.
