### 2016-08-02 #61 @sckott
thanks @fawda123 for offering, assigned above https://github.com/ropensci/onboarding/issues/61#issuecomment-236732231 - Here's instructions for review https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md and examples of good reviews: https://github.com/ropensci/onboarding/issues/22#issuecomment-164580884 and https://github.com/ropensci/onboarding/issues/20#issuecomment-148773403
