### 2016-08-15 #61 @aappling-usgs
Joseph (@jsta) and Scott (@sckott), here's my review. Marcus (@fawda123), thanks for a thoughtful review to build on.

- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README (I agree with @fawda123's suggestion)
- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and `URL`, `Maintainer` and `BugReports` fields in DESCRIPTION (just missing the contribution guidelines)

- [ ] **Functionality:** Any functional claims of the software been confirmed. (see below for some edge cases)
   and a reasonable range of inputs and conditions. All tests pass on the local machine. (but see below for some edge cases that could also be tested)

---

This package provides scriptable access to an extensive Florida water quality/quantity database, DBHYDRO. It's also noticeably faster than the browser-based interface. `dbhydroR` will enhance the usability of the database and the reproducibility of analyses of DBHYDRO data.

I agree with @fawda123 that the largest room for improvement is in the documentation. I have also made suggestions below for increasing the flexibility of the data-cleaning functions to serve several common use cases, and I've pointed out a few edge cases where the package's functionality might surprise a user.

As @fawda123 noted, it can be hard to identify the station IDs and tests you want if you don't know what the options are. There are a few ways you could helper users:
* You have several great links in the vignette that could be ported to the corresponding help files.
* The Google Earth kmz file is excellent for interactively identifying sites and timeseries. However, http://my.sfwmd.gov/KMLEXT/CUSTOMKMLS/DBHydro/DBHydroKML/DBHYDRO_KML.kmz doesn't open a browser window as a user might expect - maybe include some suggestions for how to best open the file, like on p. 56 of the DBHYDRO manual.
* If you're feeling ambitious, functions that return lists of options (sites, tests) would enhance users' ability to search sites programatically. I also see a simplicity/maintainability argument for avoiding this route.

Questions that arose on my first encounter with the package/dataset:
* What does a dbkey represent? A site, a timeseries, ...?
* How should I cite DBHYDRO and/or the `dbhydroR` package?

Other documentation thoughts
* The description of "cleaned output" in the vignette could also be copied right over to the help files for `getwd` and `gethydro` (and their internal cleaning functions if those continue to be exported).
* Contribution guidelines could be added to the README or a CONTRIBUTING file.

#### Build/install

* I had no trouble installing/building the package from CRAN, GitHub, or a local clone of the repository.
* The GitHub repository name is `dbhydro` while the package name is `dbhydroR`. You might consider switching to the same name in both places to make it ever-so-slightly easier to find the package on GitHub. (E.g., my first try was `devtools::install_github('SFWMD/dbhydroR')` because I knew the package name and only glanced at the URL.)

#### Examples/tests

* I encountered the same error as @fawda123 on running the example for `cleanhydro`:
```r
> devtools::run_examples(run=FALSE)
...
> cleanhydro(gethydro(dbkey = "15081", date_min = "2013-01-01", date_max = "2013-02-02"))
Error: is.response(x) is not TRUE
```
* `devtools::check()` runs cleanly with 0 errors, 0 warnings, and 0 notes.

#### Possibly unnecessary files & code

- The NEWS file has `output: pdf_document` at the top. Are you using this feature?
- Do you need all of the files in dbhydro/vignettes? DFlowInterpR.bib~, Thumbs.db, and Untitled Document~ (and maybe others) might not be getting used.

#### Functions

##### getwq + gethydro

These functions have a clean, intuitive interface. A few thoughts:
* It would be helpful to more thoroughly explain `dbkey` in the documentation for `gethydro`.
* Is it / could it be possible to pass `dbkey` into `getwq`, as it is for `gethydro`? One dbkey might be easier than site+test for some programming applications, and it looks like water quality data do have dbkeys, e.g.:
```r
> getdbkey(stationid = "C111%", stat = 'MEAN', category = "WQ")
Search results for 'C111% WQ'
  Dbkey    Group Data Type Freq Recorder  Start Date    End Date
1 38097 C111WETL      H2OT   DA     ???? 04-OCT-2003 10-OCT-2011
2 38104 C111WETL     SALIN   DA     ???? 04-OCT-2003 10-OCT-2011
3 38100 C111WETL     SCOND   DA     ???? 04-OCT-2003 10-OCT-2011
```
* Two things only occur in `getwq` if `raw==FALSE` which I expected to be conditional: (1) the message "No data found" doesn't occur even if there are no data, and (2) MDL handling doesn't occur even if I specify `mdl_handling`. My favorite solution would be to make these things happen regardless of the value of `raw`, but another option would be to document them prominently in the `?getwq` help file.
```r
> getwq("FLAB01", "2014-09-14", "2014-09-18", "NITRATE+NITRITE-N", raw = TRUE)$Value
[1] -0.01
> getwq("FLAB01", "2014-09-14", "2014-09-18", "NITRATE+NITRITE-N", raw = TRUE, mdl_handling='half')$Value
[1] -0.01
```

##### getdbkey

Though not advertised in the README, the `getdbkey` function looks like a useful way to identify datasets to download. With that in mind, I have a few suggestions that are mostly illustrated by this example:
```r
> getdbkey(stationid = "C111%", stat = 'MEAN', category = "WQ", detail.level="full")
Search results for 'C111% WQ'
  Get Data Dbkey \nStation\n    Group Data Type Freq Stat Recorder Agency
1          38097  C111WETLND C111WETL      H2OT   DA MEAN     ????   USGS
2          38104  C111WETLND C111WETL     SALIN   DA MEAN     ????   USGS
3          38100  C111WETLND C111WETL     SCOND   DA MEAN     ????   USGS
   Start Date    End Date Strata County Op Num Latitude Longitude    X Coord
1 04-OCT-2003 10-OCT-2011      0    DAD      Â   251740    803112 814755.394
2 04-OCT-2003 10-OCT-2011      0    DAD      Â   251740    803112 814755.394
3 04-OCT-2003 10-OCT-2011      0    DAD      Â   251740    803112 814755.394
     Y Coord   Basin Struct Section Township Range
1 349549.418 C111_CO      Â      15       59    38
2 349549.418 C111_CO      Â      15       59    38
3 349549.418 C111_CO      Â      15       59    38
```
* Calls with `detail.level='full'` return a column named `"\nStation\n"`. I'd prefer `"Station"` if there's no reason to stick with the extra newlines.
* The values of Latitude and Longitude are unconventionally formatted. Would you consider converting them to decimal degrees?
* Are the values in the `"Struct"` column being parsed correctly? I don't know what to make of `Â`. It seems to appear in other columns, too, in other queries.
* Does anything useful ever appear in the `"Get Data"` column?
* Consider using `stringsAsFactors=FALSE` so that `Station`, `Dbkey`, etc. come back as strings.
* When I try to get a more complete list of sites/tests to choose from, it seems that I'm limited to 100 dbkeys. This limitation should probably be more clearly documented and/or circumvented.
http://my.sfwmd.gov/dbhydroplsql/show_dbkey_info.show_station_info?v_station=I%25
```r
# we'd expect >100 dbkeys in this query
> nrow(getdbkey(category = "GW", detail.level = "full"))
Search results for 'NA GW'
[1] 100
# because we easily get >100 sites combined in these two queries:
> dba <- getdbkey(stationid = 'A%', category = "GW", detail.level = "full")
Search results for 'A% GW'
> dbi <- getdbkey(stationid = 'I%', category = "GW", detail.level = "full")
Search results for 'I% GW'
> nrow(dba) + nrow(dbi)
[1] 134
```

##### cleanwq + cleanhydro

* In `cleanwq` at https://github.com/SFWMD/dbhydro/blob/master/R/dbhydro_clean.R#L27 and in `cleanhydro` at https://github.com/SFWMD/dbhydro/blob/master/R/dbhydro_clean.R#L98, it looks like the timezone a user gets back will depend on the timezone where he/she is located, though probably all the data are in Eastern time, right? This is especially problematic if the data are not stored with daylight savings time or with a different daylight savings time than that of the user's locale. For example, running the following query in Arizona, I get:
```r
> h <- gethydro(dbkey = "IY639", date_min = "2009-01-30", date_max = "2009-01-31")
1: In cleanhydro(res) : Column headings missing. Guessing...
2: In cleanhydro(res) : Instantaneous data detected
> head(h$date)
[1] "2009-01-30 00:40:00 MST" "2009-01-30 00:55:00 MST" "2009-01-30 01:10:00 MST"
[4] "2009-01-30 01:25:00 MST" "2009-01-30 01:40:00 MST" "2009-01-30 01:55:00 MST"
```
* In `cleanhydro`, why is "Instantaneous data detected" a warning rather than a message? Is it possible for instantaneous data to make trouble?
* At first glance, I agreed with @fawda123 that it would make sense not to export the `cleanwq` and `cleanhydro` functions because they are wrapped by `getwd` and `gethydro`. After more thought, though, I decided that I might often want to retain the `Station.ID` column and the long format (e.g., for `dplyr`-style analyses of multiple sites or tests) but might also want to specify the `mdl_handling` and/or get back parsed timestamps. I see no one-line option for these use cases, and it makes me wonder if the cleaning functionality could be split into things you'd almost always want (parsed timestamps, connecting metadata to headers, maybe move the Value column closer to the left) and things you might want to specify in various ways (MDL handling, selecting specific columns, switching between wide and long, renaming the value column). The former functionality could be the default in getwd and gethydro while the latter could live in (a) exported functions that you'd call separately or (b) new arguments to `getwd` and `gethydro`.
* If you decide that `cleanwq` and `cleanhydro` should continue to be exported functions, they should be used in examples in the README and vignette.

#### Signing off

This package is already a nice contribution - thanks for developing it. Let me know if you'd like to clarify/discuss any of the above points. Best, Alison (@aappling-usgs)
