### 2016-08-30 #61 @sckott
@jsta Thanks for responding to reviews and making changes.

@fawda123 @aappling-usgs There's a few questions for you, see his responses. Also, any more thoughts, or are you happy?
