### 2016-09-01 #61 @fawda123
@jsta Looks like you've done a good job addressing my comments, nice work.  I'm responding below to your specific questions and some follow-up points after having some time to stew on the review.

About the wild-cards, I guess I was wondering if you could have a few more examples in the vignette of how these can be used to get info on several stations in the case where a user has some knowledge about the station naming convention but specific names are unknown.  For example, all stations that start with 'A' or 'B', etc.  Are there other characters besides '%' that can help with retrieval (e.g., regex matching)?  It's not all that important, the additional links in the vignette are very helpful.

The link for the `test_name` argument in `getwq` documentation is really helpful. Maybe also add the station-search utility [link ](http://my.sfwmd.gov/dbhydroplsql/water_quality_data.show_group_station_characters) under the `station_id` argument description.

Minor typo, 'hydrolgic' in the README description.

I'll try clarify my comment on the QA field blanks, which might be related to my misunderstanding of the data.  For an instance of using `getwq` to retrieve multiple stations, would there not be separate fields for each station?  Removing rows for fields that apply to one station would remove valid data for a station wasn't flagged if the data were in wide format, yes?  Just wanted to clarify that the cleaning function isn't removing valid data.

I also ask you to reconsider changing the name of `getwq` based on my previous comment.  As proof of point, @aaplling-usgs has used `getwd` by accident a few times.
