### 2016-09-01 #61 @aappling-usgs
Hah - true, I did end up typing `getwd` a bunch of times in my comments accidentally, didn't I? It's hard to type `getwq` when the habit is in your fingers.

@jsta, a quick glance suggests you've made some great improvements. I've been out of town and will need a few more days to catch up and reply properly.
