### 2016-09-01 #61 @jsta
@fawda123
* I believe that `%` is the only wildcard operation that is currently supported by the DBHYDRO database backend (at least it is the only one that is documented).
* Great suggestion on adding a link to the station-search utility!
* Typo fixed.
* No valid data is being removed. If `mdl_handling` is specified, flagged data is removed prior to conversion to wide format.
* You make a good argument for changing the `getwq` name. I just needed to learn what is the best way to deprecate/alias a function name so as not to break backwards compatibility. My plan is to follow the accepted answer in [this stackoverflow thread](http://stackoverflow.com/a/10145627/3362993).
