### 2016-09-02 #61 @jsta
I went ahead and refactored to underscored function names. Maybe that will break the `getwd` muscle memory :question:
