### 2016-09-02 #61 @fawda123
@jsta The underscore works better, good idea.

@sckott I'm happy with the changes from my end, this package is good to go pending comments from @aappling-usgs
