### 2016-09-03 #61 @aappling-usgs
@jsta - yeah, tons of great changes, and I'm satisfied as-is. I agree that there are several ways to address the division between getting and cleaning, and I like what you've chosen. I also like the new underscores.

I'm not particularly worried about it, but the `Â` symbols in the `Num` and sometimes also `Struct` columns are still appearing for me, even after updating to the most recent version. Maybe it's just how my system is set up, and it does seem to vary by query - the query you showed gives `'SPIL'` in the `Struct` column but still gives `Â` in the `Num` column, while a different query gives `Â`s in both.
```r
> get_dbkey(category = "SW", stationid = "ALLIGAT", param = "STG", detail.level = "full")
Search results for 'ALLIGAT SW'
  Get Data Dbkey Station   Group Data Type Freq Stat Recorder Agency  Start Date    End Date Strata County Op Num
1          00875 ALLIGAT ALLIGAT       STG   DA MEAN     ????   USGS 05-OCT-1979 06-DEC-1979      0    CHA      Â
  Latitude Longitude    X Coord    Y Coord    Basin Struct Section Township Range
1 26.34222  -81.8303 338357.138 932249.714 CHARLOTT   SPIL      23       41    23
> get_dbkey(stationid = "C111%", stat = 'MEAN', category = "WQ", detail.level="full")
Search results for 'C111% WQ'
  Get Data Dbkey    Station    Group Data Type Freq Stat Recorder Agency  Start Date    End Date Strata County Op Num
1          38097 C111WETLND C111WETL      H2OT   DA MEAN     ????   USGS 04-OCT-2003 10-OCT-2011      0    DAD      Â
2          38104 C111WETLND C111WETL     SALIN   DA MEAN     ????   USGS 04-OCT-2003 10-OCT-2011      0    DAD      Â
3          38100 C111WETLND C111WETL     SCOND   DA MEAN     ????   USGS 04-OCT-2003 10-OCT-2011      0    DAD      Â
  Latitude Longitude    X Coord    Y Coord   Basin Struct Section Township Range
1    25.74   -80.112 814755.394 349549.418 C111_CO      Â      15       59    38
2    25.74   -80.112 814755.394 349549.418 C111_CO      Â      15       59    38
3    25.74   -80.112 814755.394 349549.418 C111_CO      Â      15       59    38
```
