### 2016-09-03 #61 @jsta
@aappling-usgs My guess is that it is some sort of encoding issue. I've just pushed a commit that specifies the encoding of `httr::content()`. Does this fix the issue?
