### 2016-09-05 #61 @jsta
Ok, I think I figured it out. Those two columns were being read as `native` encoding despite setting encoding to `UTF-8` in the initial pull. My latest commit should fix (https://github.com/SFWMD/dbhydroR/commit/2125b25e9436c88ebb0a1bea28f8f373b0722177).

For future reference, I used `stringi::stri_enc_mark()` to check encoding. This [r-help thread](https://stat.ethz.ch/pipermail/r-help/2011-May/277209.html) describes the issue but it seems that my "fix" is not documented anywhere.
