### 2016-09-06 #61 @aappling-usgs
That's it!
```r
> library(dbhydroR)
> get_dbkey(category = "SW", stationid = "ALLIGAT", param = "STG", detail.level = "full")
Search results for 'ALLIGAT SW'
  Dbkey Station   Group Data Type Freq Stat Recorder Agency  Start Date    End Date Strata County Op Num Latitude
1 00875 ALLIGAT ALLIGAT       STG   DA MEAN     ????   USGS 05-OCT-1979 06-DEC-1979      0    CHA        26.34222
  Longitude    X Coord    Y Coord    Basin Struct Section Township Range
1  -81.8303 338357.138 932249.714 CHARLOTT   SPIL      23       41    23
> get_dbkey(stationid = "C111%", stat = 'MEAN', category = "WQ", detail.level="full")
Search results for 'C111% WQ'
  Dbkey    Station    Group Data Type Freq Stat Recorder Agency  Start Date    End Date Strata County Op Num Latitude
1 38097 C111WETLND C111WETL      H2OT   DA MEAN     ????   USGS 04-OCT-2003 10-OCT-2011      0    DAD           25.74
2 38104 C111WETLND C111WETL     SALIN   DA MEAN     ????   USGS 04-OCT-2003 10-OCT-2011      0    DAD           25.74
3 38100 C111WETLND C111WETL     SCOND   DA MEAN     ????   USGS 04-OCT-2003 10-OCT-2011      0    DAD           25.74
  Longitude    X Coord    Y Coord   Basin Struct Section Township Range
1   -80.112 814755.394 349549.418 C111_CO             15       59    38
2   -80.112 814755.394 349549.418 C111_CO             15       59    38
3   -80.112 814755.394 349549.418 C111_CO             15       59    38
```
