### 2016-09-06 #61 @sckott
thanks @aappling-usgs and @fawda123 for your reviews!

@jsta Just looking over it before approving, a few questions:

In e.g.,

```r
get_wq(station_id = "FLAB08", date_min = "2011-03-01",
    date_max = "2012-05-01", test_name = "CHLOROPHYLLA-SALINE")
#>          date FLAB08_CHLOROPHYLLA-SALINE_ug/L
#> 1  2011-03-21                           0.196
#> 2  2011-04-11                           0.276
#> 3  2011-05-09                           0.328
#> 4  2011-06-07                           0.329
#> 5  2011-07-05                           0.402
#> 6  2011-08-08                           0.907
#> 7  2011-09-06                           0.887
#> 8  2011-10-06                           0.573
#> 9  2011-12-08                           0.369
#> 10 2012-02-02                           0.260
#> 11 2012-04-05                           0.509
```

Is the 2nd column name a combination of the station name and the variable? If so, I'd prefer that this is changed to be in more of tidy format, with station as  it's own column, and variable as its own column. At least have a function to get it in that format easily would be nice.
