### 2016-09-06 #61 @jsta
Thanks for your question @sckott.

There is already a method for returning station name and variable in separate columns.

```r
get_wq(station_id = "FLAB08", date_min = "2011-03-01",
    date_max = "2012-05-01", test_name = "CHLOROPHYLLA-SALINE",
    raw = TRUE)[,c("date", "Test.Name", "Units", "Value")]

>  date Test.Name Units Value
> 1  2011-03-21 CHLOROPHYLLA-SALINE  ug/L 0.196
> 3  2011-04-11 CHLOROPHYLLA-SALINE  ug/L 0.276
> 5  2011-05-09 CHLOROPHYLLA-SALINE  ug/L 0.328
> 6  2011-06-07 CHLOROPHYLLA-SALINE  ug/L 0.329
> 7  2011-07-05 CHLOROPHYLLA-SALINE  ug/L 0.402
> 9  2011-08-08 CHLOROPHYLLA-SALINE  ug/L 0.907
> 10 2011-09-06 CHLOROPHYLLA-SALINE  ug/L 0.887
> 11 2011-10-06 CHLOROPHYLLA-SALINE  ug/L 0.573
> 12 2011-12-08 CHLOROPHYLLA-SALINE  ug/L 0.369
> 13 2012-02-02 CHLOROPHYLLA-SALINE  ug/L 0.260
> 14 2012-04-05 CHLOROPHYLLA-SALINE  ug/L 0.509
```
The reason why `raw` is an argument rather than the default is because my initial feedback on the package was from users who appreciated receiving output in non-tidy format for simple plotting and Excel import.

Also, it was valuable for users to do things like:
```r
plot(get_wq(station_id = "FLAB08", date_min = "2011-03-01",
    date_max = "2012-05-01", test_name = "CHLOROPHYLLA-SALINE"))
```

 Do you think I should do away with the `raw` argument and make a separate function?
