### 2016-09-06 #61 @sckott
@jsta Thanks. I think it's okay to leave as is, the `raw` parameter takes care of it I think.
