### 2016-09-28 #61 @sckott
thanks again for the reviews @aappling-usgs and @fawda123   🚀 ✏️ 📋

@jsta  - There's a few things I'd like changed before approving:

-   Please use a `README.Rmd` and build your `README.md` from that, instead of having to manually create `README.md`, which I assume you're doing (correct me if I'm wrong)
-  I noticed one thing to deal with after running `goodpractice::gp()` at some point (no need to do prior to transferring though):
    * Avoid long code and documentation lines. Stick to 80 characters or less for each line. Looks like your code does this, but docs should too.
- I suggest moving your code to do HTTP requests into helper functions that are general enough that you can use them in your various functions that need to do HTTP requests, e.g,

```r
dbh_GET <- function(url, ...) {
    res <- httr::GET(url, ...) # any params can be passed to GET() when calling dbh_GET()
    stop_for_status(res) # make sure to check if the http status is good or not
    httr::content(res, "text", encoding = "UTF-8") # parse to text
}
```

Then use `dbh_GET()` in your functions.

------------

Then after that, the steps to move to ropensci are:

-   Add the rOpenSci footer to the README
-   Transfer the repo to ropenscilabs - I've created a team in that org. account, and you should have received an email about, after you accept the invitation, go to Settings->Transfer Ownership
-   After transferring to ropenscilabs, travis builds should work automatically under the new github account. but you do need to update the link in the badge in your readme
-  Update all links to github repo to replace `jsta` with `ropenscilabs`
