### 2016-09-29 #61 @jsta
Ok, I think I've addressed all the items you've listed under pre-approval. Should I move forward with transferring? @sckott
