### 2016-09-29 #61 @sckott
thanks for the change. Yes, do move forward
